requirements.txt
python_telegram_bot
ffmpeg_python